+++ src/components/Logo/Logo.js
@@ -1,21 +1,18 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
 
 import config from '../../config';
-import IconLogo from './IconLogo';
-import LogoImage from './saunatime-logo.png';
-import css from './Logo.module.css';
+import LogoImageMobile from './sneakertime-logo-mobile.png';
+import LogoImage from './sneakertime-logo.png';
 
 const Logo = props => {
   const { className, format, ...rest } = props;
-  const mobileClasses = classNames(css.logoMobile, className);
 
   if (format === 'desktop') {
     return <img className={className} src={LogoImage} alt={config.siteTitle} {...rest} />;
   }
 
-  return <IconLogo className={mobileClasses} {...rest} />;
+  return <img className={className} src={LogoImageMobile} alt={config.siteTitle} {...rest} />;
 };
 
 const { oneOf, string } = PropTypes;