+++ src/components/Topbar/TopbarDesktop/TopbarDesktop.test.js
@@ -6,7 +6,7 @@ import TopbarDesktop from './TopbarDesktop';
 const noop = () => null;
 
 describe('TopbarDesktop', () => {
-  it('data matches snapshot', () => {
+  it('data with location search matches snapshot', () => {
     window.google = { maps: {} };
     const topbarProps = {
       isAuthenticated: true,
@@ -15,6 +15,23 @@ describe('TopbarDesktop', () => {
       onSearchSubmit: noop,
       intl: fakeIntl,
       onLogout: noop,
+      appConfig: { mainSearchType: 'location' },
+    };
+    const tree = renderDeep(<TopbarDesktop {...topbarProps} />);
+    delete window.google;
+    expect(tree).toMatchSnapshot();
+  });
+
+  it('data with keywords search matches snapshot', () => {
+    window.google = { maps: {} };
+    const topbarProps = {
+      isAuthenticated: true,
+      currentUserHasListings: true,
+      name: 'John Doe',
+      onSearchSubmit: noop,
+      intl: fakeIntl,
+      onLogout: noop,
+      appConfig: { mainSearchType: 'keywords' },
     };
     const tree = renderDeep(<TopbarDesktop {...topbarProps} />);
     delete window.google;