+++ src/config/config.js
@@ -19,11 +19,16 @@ const i18n = {
   firstDayOfWeek: 0,
 };
 
+// Main search used in Topbar.
+// This can be either 'keywords' or 'location'.
+const mainSearchType = 'keywords';
+
 // Should search results be ordered by distance to origin.
-// NOTE: If this is set to true add parameter 'origin' to every location in default-location-searches.js
-// Without the 'origin' parameter, search will not work correctly
-// NOTE: Keyword search and ordering search results by distance can't be used at the same time. You can turn keyword
-// search off by changing the keywordFilterConfig parameter active to false in marketplace-custom-config.js
+// NOTE 1: This doesn't affect if the main search type is 'keywords'
+// NOTE 2: If this is set to true add parameter 'origin' to every location in default-location-searches.js
+//         Without the 'origin' parameter, search will not work correctly
+// NOTE 3: Keyword search and ordering search results by distance can't be used at the same time. You can turn keyword
+//         search off by changing the keywordFilterConfig parameter active to false in marketplace-custom-config.js
 const sortSearchByDistance = false;
 
 // API supports custom processes to be used in booking process.
@@ -203,6 +208,7 @@ const config = {
     baseUrl: sdkBaseUrl,
     transitVerbose: sdkTransitVerbose,
   },
+  mainSearchType,
   sortSearchByDistance,
   currency,
   currencyConfig,